+++ mycodo/outputs/on_off_python.py
@@ -71,7 +71,7 @@ def execute_at_modification(
         custom_options_dict_postsave,
         custom_options_channels_dict_postsave):
     """
-    Function to run when the Output is saved to evaluate the Python 3 code using pylint3
+    Function to run when the Output is saved to evaluate the Python 3 code using pylint
     :param messages: dict of info, warning, error, success messages as well as other variables
     :param mod_input: The WTForms object containing the form data submitted by the web GUI
     :param request_form: The custom_options form input data (if it exists)
@@ -112,7 +112,7 @@ def execute_at_modification(
         cmd_test = 'mkdir -p /var/mycodo-root/.pylint.d && ' \
                    'export PYTHONPATH=$PYTHONPATH:/var/mycodo-root && ' \
                    'export PYLINTHOME=/var/mycodo-root/.pylint.d && ' \
-                   'pylint3 -d I,W0621,C0103,C0111,C0301,C0327,C0410,C0413 {path}'.format(
+                   'pylint -d I,W0621,C0103,C0111,C0301,C0327,C0410,C0413 {path}'.format(
                        path=file_run)
         cmd_out, cmd_error, cmd_status = cmd_output(cmd_test)
         pylint_message = Markup(
@@ -170,6 +170,10 @@ OUTPUT_INFORMATION = {
         'button_send_duration'
     ],
     'options_disabled': ['interface'],
+
+    'dependencies_module': [
+        ('apt', 'pylint', 'pylint'),
+    ],
 
     'interfaces': ['PYTHON'],