+++ mycodo/outputs/pwm_python.py
@@ -25,6 +25,9 @@ from mycodo.utils.system_pi import cmd_output
 from mycodo.utils.system_pi import return_measurement_info
 from mycodo.utils.system_pi import set_user_grp
 
+import logging
+logger = logging.getLogger("mycodo.influxdb")
+
 
 def generate_code(code_pwm, unique_id):
     pre_statement_run = f"""import os
@@ -71,7 +74,7 @@ def execute_at_modification(
         custom_options_dict_postsave,
         custom_options_channels_dict_postsave):
     """
-    Function to run when the Output is saved to evaluate the Python 3 code using pylint3
+    Function to run when the Output is saved to evaluate the Python 3 code using pylint
     :param messages: dict of info, warning, error, success messages as well as other variables
     :param mod_input: The WTForms object containing the form data submitted by the web GUI
     :param request_form: The custom_options form input data (if it exists)
@@ -111,7 +114,7 @@ def execute_at_modification(
         cmd_test = 'mkdir -p /var/mycodo-root/.pylint.d && ' \
                    'export PYTHONPATH=$PYTHONPATH:/var/mycodo-root && ' \
                    'export PYLINTHOME=/var/mycodo-root/.pylint.d && ' \
-                   'pylint3 -d I,W0621,C0103,C0111,C0301,C0327,C0410,C0413 {path}'.format(
+                   'pylint -d I,W0621,C0103,C0111,C0301,C0327,C0410,C0413 {path}'.format(
                        path=file_run)
         cmd_out, cmd_error, cmd_status = cmd_output(cmd_test)
         pylint_message = Markup(
@@ -156,7 +159,7 @@ channels_dict = {
 # Output information
 OUTPUT_INFORMATION = {
     'output_name_unique': 'python_pwm',
-    'output_name': "Python Code: {}".format(lazy_gettext('PWM')),
+    'output_name': "Python 3 Code: {}".format(lazy_gettext('PWM')),
     'measurements_dict': measurements_dict,
     'channels_dict': channels_dict,
     'execute_at_modification': execute_at_modification,
@@ -169,6 +172,10 @@ OUTPUT_INFORMATION = {
         'button_send_duty_cycle'
     ],
     'options_disabled': ['interface'],
+
+    'dependencies_module': [
+        ('apt', 'pylint', 'pylint'),
+    ],
 
     'interfaces': ['PYTHON'],
 
@@ -183,8 +190,8 @@ log_string = "ID: {id}: {dc} % Duty Cycle".format(
 self.logger.info(log_string)""",
             'required': True,
             'col_width': 12,
-            'name': lazy_gettext('Bash Command'),
-            'phrase': 'Command to execute to set the PWM duty cycle (%)'
+            'name': 'Python 3 Code',
+            'phrase': 'Python code to execute to set the PWM duty cycle (%)'
         },
         {
             'id': 'linux_command_user',