+++ mycodo/utils/code_verification.py
@@ -53,7 +53,7 @@ def test_python_code(python_code_run, filename):
         cmd_test = 'mkdir -p /var/mycodo-root/.pylint.d && ' \
                    'export PYTHONPATH=$PYTHONPATH:/var/mycodo-root && ' \
                    'export PYLINTHOME=/var/mycodo-root/.pylint.d && ' \
-                   'pylint3 -d I,W0621,C0103,C0111,C0301,C0327,C0410,C0413,R0201,R0903,W0201,W0612 {path}'.format(
+                   'pylint -d I,W0621,C0103,C0111,C0301,C0327,C0410,C0413,R0201,R0903,W0201,W0612 {path}'.format(
                        path=file_run)
         cmd_out, _, cmd_status = cmd_output(cmd_test)